[proliferate] split workspace browser panel
