+++ desktop/src/components/workspace/browser/BrowserPanelTypes.tsx
@@ -1,0 +1,1 @@
+export type FrameStatus = "idle" | "loading" | "loaded" | "blocked";
